backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
psycopg2-binary>=2.9
gunicorn>=21.2
Pillow>=11.0
whitenoise>=6.6
djangorestframework-simplejwt>=5.2
pytest>=8.0
pytest-django>=4.11
pytest-cov>=6.1
